[backend] finalisation ready
