+++ fc-community-backend/apps/api/package.json
@@ -25,7 +25,7 @@
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "axios": "^1.6.0",
-    "ethers": "^6.8.0",
+    "ethers": "^6.10.0",
     "reflect-metadata": "^0.1.13",
     "replicate": "^0.22.0",
     "rxjs": "^7.8.1"
